Add types and clean up logic
