+++ src/lib/@types/plugins.ts
@@ -1,0 +1,16 @@
+export interface PluginListResponse {
+  command: 'list'
+  plugins: Plugin[]
+}
+
+export type Plugin = {
+  name: string
+  active: boolean
+  dynamic: boolean
+}
+
+export type ClbossStatus = {
+  incoming_capacity_swapper: {
+    status: string
+  }
+}
