+++ src/lib/wallets/coreln/clboss.ts
@@ -1,3 +1,4 @@
+import type { ClbossStatus } from '$lib/@types/plugins.js'
 import type { ClbossInterface } from '../interfaces.js'
 import type { CorelnConnectionInterface } from './types.js'
 
@@ -8,13 +9,11 @@ class Clboss implements ClbossInterface {
     this.connection = connection
   }
 
-  async get(): Promise<[]> {
-    console.log('clboss-status called!')
-    const result = await this.connection.rpc({
+  async get(): Promise<ClbossStatus> {
+    const result = (await this.connection.rpc({
       method: 'clboss-status'
-    })
-    console.log(`result = `, result)
-    return result // TODO add GetClbossStatusResponse type
+    })) as ClbossStatus
+    return result
   }
 }
